requirements.txt
flask
flask-cors
python-dotenv
torch
transformers
google-generativeai
